Adapted Angular to Modular Pattern
