+++ types/angular.ts
@@ -1,0 +1,2 @@
+// Structural type — no @angular/core import needed
+export type AngularComponent<T = unknown> = new (...args: any[]) => T;
