+++ types/index.ts
@@ -1,3 +1,4 @@
+export * from './angular';
 export * from './build';
 export * from './cli';
 export * from './client';
